feat: move sqs_collect_messages to helperfunctions
