+++ tests/aws/services/events/test_events_inputs.py
@@ -7,8 +7,11 @@ from botocore.client import Config
 
 from localstack.testing.pytest import markers
 from localstack.utils.strings import short_uid
-from tests.aws.services.events.conftest import sqs_collect_messages
-from tests.aws.services.events.helper_functions import is_old_provider, is_v2_provider
+from tests.aws.services.events.helper_functions import (
+    is_old_provider,
+    is_v2_provider,
+    sqs_collect_messages,
+)
 from tests.aws.services.events.test_events import EVENT_DETAIL, TEST_EVENT_PATTERN
 
 EVENT_DETAIL_DUPLICATED_KEY = {
