+++ tests/aws/services/events/test_events_integrations.py
@@ -13,8 +13,8 @@ from localstack.utils.aws import arns, resources
 from localstack.utils.strings import short_uid
 from localstack.utils.sync import retry
 from localstack.utils.testutil import check_expected_lambda_log_events_length
-from tests.aws.services.events.conftest import assert_valid_event, sqs_collect_messages
-from tests.aws.services.events.helper_functions import is_v2_provider
+from tests.aws.services.events.conftest import assert_valid_event
+from tests.aws.services.events.helper_functions import is_v2_provider, sqs_collect_messages
 from tests.aws.services.events.test_events import EVENT_DETAIL, TEST_EVENT_PATTERN
 from tests.aws.services.lambda_.test_lambda import TEST_LAMBDA_PYTHON_ECHO
